fix(SDKConfigValues+Validations) : Change the SDKConfig struct and move the validations foor the correct place
