+++ pkg/controller/mobilesecurityserviceapp/validations.go
@@ -1,0 +1,23 @@
+package mobilesecurityserviceapp
+
+import (
+	mobilesecurityservicev1alpha1 "github.com/aerogear/mobile-security-service-operator/pkg/apis/mobilesecurityservice/v1alpha1"
+	"github.com/go-logr/logr"
+)
+
+//Check if the mandatory specs are filled
+func hasMandatorySpecs(instance *mobilesecurityservicev1alpha1.MobileSecurityServiceApp, serviceInstance *mobilesecurityservicev1alpha1.MobileSecurityService, reqLogger logr.Logger) bool {
+	//Check if the appId was added in the CR
+	if len(instance.Spec.AppId) < 1 {
+		reqLogger.Info("AppID was not found. Check the App CR configuration.")
+		return false
+	}
+
+	//Check if the appId was added in the CR
+	if len(instance.Spec.AppId) < 1 {
+		reqLogger.Info("AppName was not found. Check the App CR configuration.")
+		return false
+	}
+
+	return true
+}
