+++ pkg/models/sdk_config.go
@@ -5,28 +5,29 @@ import (
 )
 
 type SDKConfig struct {
-	Version   string             `json:"version"`
-	Name      string             `json:"name"`
-	Namespace string             `json:"namespace"`
-	ClientID  string             `json:"clientId"`
-	Services  []SDKConfigService `json:"services,omitempty"`
+	ID            string        `json:"id"`
+	Name          string        `json:"name"`
+	Type          string        `json:"type"`
+	URL           string        `json:"url"`
+	ConfigService ConfigService `json:"config,omitempty"`
 }
 
 func NewSDKConfig(m *mobilesecurityservicev1alpha1.MobileSecurityServiceApp, serviceUrl string) *SDKConfig {
-
 	cfg := new(SDKConfig)
-	cfg.Version = "1.0.0"
+	cfg.ID = m.Spec.AppId
 	cfg.Name = m.Spec.AppName
-	cfg.ClientID = m.Spec.AppId
-	cfg.Namespace = m.Namespace
-	cfg.Services = getServices(m, serviceUrl)
+	cfg.Type = m.Name
+	cfg.URL = serviceUrl
+	cfg.ConfigService = *NewConfigService(serviceUrl)
 	return cfg
 }
 
-//return the Service data for the SDK ConfigMap
-func getServices(m *mobilesecurityservicev1alpha1.MobileSecurityServiceApp, serviceUrl string) []SDKConfigService {
-	service := *NewSDKConfigServices(m, serviceUrl)
-	res := []SDKConfigService{}
-	res = append(res, service)
-	return res
+type ConfigService struct {
+	MobileSecurityServiceURL string `json:"mobile-security-server-url"`
+}
+
+func NewConfigService(url string) *ConfigService {
+	service := new(ConfigService)
+	service.MobileSecurityServiceURL = url
+	return service
 }
